refactor(insights): move data struct to its own package
